Hook up stdin to apt-get properly
If apt-get/dpkg throws up any prompts to the user (e.g. asking to overwrite
existing collectd config files), the user was previously unable to answer since
standard input was not propertly directed to the forked apt-get process
(probably due to how we are piping stdin from curl originally).  This forces
stdin to the controlling tty terminal of the bash script.